requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
newspaper3k==0.2.8
transformers
torch
scipy
numpy
pandas
python-dotenv
torch
IPython
soundfile
deep_translator
pydub
bertopic
flask
sentence_transformers 
streamlit
lxml_html_clean
google-generativeai
lxml[html_clean]

